Feat: git pull origin develop 후 오류제거

+++ src/main/java/com/example/approval24/controller/AuthorityController.java
@@ -11,7 +11,12 @@ import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.example.approval24.domain.AuthorityDTO;
@@ -106,7 +111,7 @@ public class AuthorityController {
         model.addAttribute("authorityDepartments", authorityDepartments);
 
         // 미할당 부서 목록 조회
-        Set<Integer> assignedDeptIds = authorityDepartments.stream()
+        Set<Long> assignedDeptIds = authorityDepartments.stream()
                 .map(DeptDTO::getDeptId)
                 .collect(Collectors.toSet());
